requirements.txt
streamlit
openai>=1.14.0
python-dotenv

